+++ apps/website/components/templates/dao-profile/index.ts
@@ -1,1 +1,2 @@
 export * from './dao-profile';
+export * from './context';
